crawler_backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
fake-useragent==1.4.0
python-dotenv==1.0.0
loguru==0.7.2 